implementing exercise 4.21* step9

+++ controllers/blog.js
@@ -39,8 +39,25 @@ blogRouter.put('/:id', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({error: 'token missing or invalid'})
+  }
+
+  const user = await User.findById(decodedToken.id)
+
   const {id} = request.params
+  // check if blog will be deleted with by a wrong user
+  const blogToDelete = await Blog.findById(id)
+  if (blogToDelete.author.toString() !== user.id.toString()) {
+    return response.status(401).json({error: 'wrong author'})
+  }
+
   await Blog.findByIdAndRemove(id)
+
+  user.blogs = user.blogs.filter((blog) => blog.id !== id)
+  await user.save()
+
   response.status(204).end()
 })
